+++ libraries/common/src/main/java/androidx/media3/common/Player.java
@@ -1018,15 +1018,6 @@ public interface Player {
     default void onMaxSeekToPreviousPositionChanged(long maxSeekToPreviousPositionMs) {}
 
     /**
-     * @deprecated Seeks are processed without delay. Listen to {@link
-     *     #onPositionDiscontinuity(PositionInfo, PositionInfo, int)} with reason {@link
-     *     #DISCONTINUITY_REASON_SEEK} instead.
-     */
-    @Deprecated
-    @UnstableApi
-    default void onSeekProcessed() {}
-
-    /**
      * Called when the audio session ID changes.
      *
      * <p>{@link #onEvents(Player, Events)} will also be called to report this event along with