+++ libraries/common/src/test/java/androidx/media3/common/SimpleBasePlayerTest.java
@@ -1395,7 +1395,6 @@ public class SimpleBasePlayerTest {
                 /* adIndexInAdGroup= */ C.INDEX_UNSET),
             Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onMediaItemTransition(mediaItem1, Player.MEDIA_ITEM_TRANSITION_REASON_SEEK);
-    verify(listener).onSeekProcessed();
     verify(listener)
         .onEvents(
             player,
@@ -6911,7 +6910,6 @@ public class SimpleBasePlayerTest {
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
     verify(listener).onMediaItemTransition(newMediaItem, Player.MEDIA_ITEM_TRANSITION_REASON_SEEK);
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
   }
 
@@ -6962,7 +6960,6 @@ public class SimpleBasePlayerTest {
     verify(listener).onMediaItemTransition(newMediaItem, Player.MEDIA_ITEM_TRANSITION_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7022,7 +7019,6 @@ public class SimpleBasePlayerTest {
     verify(listener).onMediaItemTransition(newMediaItem, Player.MEDIA_ITEM_TRANSITION_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7074,7 +7070,6 @@ public class SimpleBasePlayerTest {
     assertThat(player.getTotalBufferedDuration()).isEqualTo(0);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7126,7 +7121,6 @@ public class SimpleBasePlayerTest {
     assertThat(player.getTotalBufferedDuration()).isEqualTo(0);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7178,7 +7172,6 @@ public class SimpleBasePlayerTest {
     assertThat(player.getTotalBufferedDuration()).isEqualTo(0);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7230,7 +7223,6 @@ public class SimpleBasePlayerTest {
     assertThat(player.getTotalBufferedDuration()).isEqualTo(7000);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7282,7 +7274,6 @@ public class SimpleBasePlayerTest {
     assertThat(player.getTotalBufferedDuration()).isEqualTo(3000);
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);
@@ -7340,7 +7331,6 @@ public class SimpleBasePlayerTest {
     verify(listener).onPositionDiscontinuity(Player.DISCONTINUITY_REASON_SEEK);
     verify(listener).onPositionDiscontinuity(any(), any(), eq(Player.DISCONTINUITY_REASON_SEEK));
     verify(listener).onMediaItemTransition(mediaItem, Player.MEDIA_ITEM_TRANSITION_REASON_SEEK);
-    verify(listener).onSeekProcessed();
     verifyNoMoreInteractions(listener);
 
     future.set(null);